Move relevant old test sections to new files
