+++ test/unit/test_bank_handler.py
@@ -17,6 +17,33 @@ class TestBankHandler(TestCase):
         # test_bank_handler.run()
         self.assertEqual(1, 1)
 
+    '''def test_init_and_name(self):
+        """Check parameters are correctly stored in the object."""
+        self.b = B2YBank(self.defaults)
+        cfe = copy(self.defaults)
+        self.assertEqual(self.b.config, cfe)
+        self.assertEqual("DEFAULT", self.b.name)'''
+
+    '''def test_get_files(self):
+        """Test it's finding the right amount of files"""
+        # if you need more tests, add sections to test.conf & specify them here
+        for section_name, num_files in [
+            ("test_num_files", 2),
+            ("test_num_files_noexist", 0),
+            ("test_num_files_extension", 0),
+            ("test_regex", 1),
+            ("test_regex_noexist", 0),
+        ]:
+
+            config = fix_conf_params(self.cp, section_name)
+            config["path"] = join(get_project_dir(), config["path"])
+            b = B2YBank(config)
+            files = b.get_files()
+            self.assertEqual(len(files), num_files)
+            # hack config to make sure we can deal with absolute paths too
+            b.config["path"] = abspath(self.test_data)
+            files = b.get_files()
+            self.assertEqual(len(files), num_files)'''
     """
     def run(self) -> list:
         transaction_reader = TransactionFileReader(
@@ -105,3 +132,31 @@ class TestBankHandler(TestCase):
         df_handler.output_csv(target_filename)
         return target_filename
     """
+    """def test_write_data(self):
+
+        Test that the right amount of files are created
+        and that the file paths end up where we expect
+        # to do - make sure file contents are what we expect
+
+        # if you need more tests, add sections to test.conf & specify them here
+        # todo: incorporate multiple-file scenarios
+        # todo: allow incremental file suffixes when files named the same
+        for section_name, num_records, fpath in [
+            ("test_record_i18n", 74, "fixed_test_raiffeisen_01.csv"),
+            ("test_record_headers", 74, "fixed_test_headers.csv"),
+        ]:
+            config = fix_conf_params(self.cp, section_name)
+            b = B2YBank(config)
+            for f in b.get_files():
+                output_data = b.read_data(f)
+                self.assertEqual(len(output_data), num_records)
+                result_file = b.write_data(f, output_data)
+                # check the file is where we expect it to be
+                expected_file = abspath(join(self.test_data, fpath))
+                self.assertTrue(exists(expected_file))
+                self.assertEqual(expected_file, result_file)
+                # todo: check actual contents are what we expect
+                os.unlink(expected_file)"""
+
+    def test_preprocess_file(self):
+        raise NotImplementedError
